* Added ability to specify a local file path or URL for a product file instead of having to upload one * Product file downloads now work with extremely large files as well

+++ application/model/ObjectFile.php
@@ -9,13 +9,13 @@ class ObjectFileException extends ApplicationException { }
  * other entities in the future
  *
  * @package application.model
- * @author Integry Systems <http://integry.com>   
+ * @author Integry Systems <http://integry.com>
  */
 class ObjectFile extends MultilingualObject
-{			
+{
 	private $sourceFilePath = false;
 	private $newFileUploaded = false;
-	
+
 	public static function defineSchema($className = __CLASS__)
 	{
 		$schema = self::getSchemaInstance($className);
@@ -24,38 +24,44 @@ class ObjectFile extends MultilingualObject
 		$schema->registerField(new ARPrimaryKeyField("ID", ARInteger::instance()));
 		$schema->registerField(new ARField("fileName", ARText::instance()));
 		$schema->registerField(new ARField("extension", ARText::instance()));
-		
+		$schema->registerField(new ARField("filePath", ARText::instance()));
 		return $schema;
-	}	
+	}
 
-	public static function getNewInstance($className, $sourceFilePath, $fileName)
+	public static function getNewInstance($className, $sourceFilePath, $fileName, $pathOrUrl = null)
 	{
 		$fileInstance = parent::getNewInstance($className);
-		$fileInstance->storeFile($sourceFilePath, $fileName);
-		
+
+		$fileInstance->storeFile($sourceFilePath, $fileName, $pathOrUrl);
+
 		return $fileInstance;
 	}
 
-	public function setBaseName($baseName) 
-	{				
-		$this->newFileUploaded = true;
-
+	public function setBaseName($baseName)
+	{
 		// write to database
 		$fileInfo = pathinfo($baseName);
-		$this->fileName->set($fileInfo['filename']);
-		$this->extension->set($fileInfo['extension']);
+		if (!empty($fileInfo['filename']))
+		{
+			$this->fileName->set($fileInfo['filename']);
+		}
+
+		if (!empty($fileInfo['extension']))
+		{
+			$this->extension->set($fileInfo['extension']);
+		}
 	}
-	
+
 	public function delete()
 	{
 		$this->deleteFile();
 		parent::delete();
 	}
-	
+
 	public function deleteFile()
 	{
 		$path = $this->getPath();
-		if(is_file($path)) 
+		if(is_file($path) && !$this->filePath->get())
 		{
 			return unlink($path);
 		}
@@ -64,23 +70,29 @@ class ObjectFile extends MultilingualObject
 			return false;
 		}
 	}
-	
-	public function storeFile($sourceFilePath, $fileName) 
+
+	public function storeFile($sourceFilePath, $fileName, $pathOrUrl = null)
 	{
-		$this->setBaseName($fileName);
+		if (!$pathOrUrl)
+		{
+			$this->newFileUploaded = true;
+		}
+
+		$this->setBaseName($fileName ? $fileName : $pathOrUrl);
 		$this->sourceFilePath = $sourceFilePath;
+		$this->filePath->set($pathOrUrl);
 	}
-	
+
 	public function save($forceOperation = false)
 	{
 		parent::save($forceOperation);
-		
+
 		if($this->newFileUploaded) $this->moveFile();
 	}
 
 	public function getPath()
 	{
-		if(!$this->isExistingRecord()) 
+		if(!$this->isExistingRecord())
 		{
 			if ($this->sourceFilePath)
 			{
@@ -91,43 +103,59 @@ class ObjectFile extends MultilingualObject
 				throw new ObjectFileException('Instance has no existing database record');
 			}
 		}
-		
+
+		if ($this->filePath->get())
+		{
+			return $this->filePath->get();
+		}
+
 		return ClassLoader::getRealPath('storage.' . strtolower(get_class($this))) . DIRECTORY_SEPARATOR . $this->getID();
 	}
-	
+
 	public function getMimeType()
 	{
 		$baseMimeTypesFile = ClassLoader::getRealPath('application.configuration.fileType.base') . '.ini';
 		$extendedMimeTypesFile = ClassLoader::getRealPath('application.configuration.fileType.extended') . '.ini';
-		
+
 		$baseTypes = is_file($baseMimeTypesFile) ? parse_ini_file($baseMimeTypesFile) : array();
 		$extendedTypes = is_file($extendedMimeTypesFile) ? parse_ini_file($extendedMimeTypesFile) : array();
-		
+
 		$baseTypes = array_merge($baseTypes, $extendedTypes);
-		
+
 		return isset($baseTypes[$this->extension->get()]) ? $baseTypes[$this->extension->get()] : 'application/octet-stream';
 	}
 
-	public function getBaseName() 
+	public function getBaseName()
 	{
 		return $this->fileName->get().'.'.$this->extension->get();
 	}
-	
+
 	public function getSize()
 	{
-		return filesize($this->getPath());
+		return $this->isLocalFile() && $this->fileExists() ? filesize($this->getPath()) : null;
 	}
-	
+
 	public function getContents()
 	{
-		return file_get_contents($this->getPath());
+		return ($this->isLocalFile() && $this->fileExists()) ? file_get_contents($this->getPath()) : '';
+	}
+
+	public function isLocalFile()
+	{
+		$path = strtolower($this->filePath->get());
+		return !((substr($path, 0, 7) == 'http://') || (substr($path, 0, 8) == 'https://'));
+	}
+
+	public function fileExists()
+	{
+		return file_exists($this->getPath());
 	}
 
 	private function moveFile()
-	{   
+	{
 		$productFileCategoryPath = ClassLoader::getRealPath('storage.' . strtolower(get_class($this)));
 		if(!is_dir($productFileCategoryPath)) mkdir($productFileCategoryPath, 0777, true);
-		
+
 		copy($this->sourceFilePath, $this->getPath());
 	}
 }
